restaurant_fare/requirements.txt
Django==1.8
django-cors-headers==1.1.0
django-rest-swagger==0.2.9
djangorestframework==3.1.1
djangorestframework-recursive==0.1.0
psycopg2==2.6
requests==2.5.3
pymongo==3.1.1
